# openapi.yaml
openapi: 3.0.3
info:
  title: Joyna API
  description: >
    API for creating, viewing, updating and sharing events within a user's network.
    Events can be forwarded by invited users to their own networks, and users can
    organize members of their network into groups.
  version: '1.0.0'
servers:
  - url: /
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List the caller's groups
      description: >
        Returns every group the caller owns, including groups with no members
        yet — unlike deriving groups from GET /network, this surfaces a group
        immediately after creation.
      operationId: listGroups
      responses:
        '200':
          description: The caller's groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags: [Groups]
      summary: Create a new group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupPayload'
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request body, or empty group name
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /groups/{id}:
    patch:
      tags: [Groups]
      summary: Update an existing group
      description: Only the group owner may update the group. Omitted fields are left unchanged.
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupPayload'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request body/id, or empty group name
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Group not found
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: A group with this name already exists
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags: [Groups]
      summary: Delete a group
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted
        '400':
          description: Invalid id
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Group not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
...
components:
  parameters:
...
    GroupId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        text/plain:
          schema:
            type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session-based authentication managed by alexedwards/scs. The cookie is set
        by /auth/register or /auth/login and must be sent on subsequent requests to
        protected routes.
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        isFavorite:
          type: boolean
      required: [id, ownerId, name, createdAt, isFavorite]
    CreateGroupPayload:
      type: object
      properties:
        name:
          type: string
      required: [name]
    UpdateGroupPayload:
      type: object
      description: All fields are optional; omitted fields are left unchanged.
      properties:
        name:
          type: string
        isFavorite:
          type: boolean
...
security:
  - sessionCookie: []